# PYTHON-4640 Improve performance of creating ObjectIds with multiple threads