PYTHON-4640.

Results before:
```
$ python bench-objectid.py         
Python: 3.12.4, PyMongo: 4.9.0.dev0
Creating 1000000 ObjectIds with 1 thread
run_threaded: 0.85s
Creating 1000000 ObjectIds across 100 threads
run_threaded: 5.25s
```
Results after:
```
$ python bench-objectid.py
Python: 3.12.4, PyMongo: 4.9.0.dev0
Creating 1000000 ObjectIds with 1 thread
run_threaded: 0.77s
Creating 1000000 ObjectIds across 100 threads
run_threaded: 0.78s
```